Updating Lab06 for swtching to WAQI API
Since http://opendata.epa.gov.tw/ will upgrade to https://data.epa.gov.tw/
and the TLS record for the new site is too large for ESP8266 MicroPython
to handle it. We decide to switch to http://waqi.info site.

The Lab06_WAQI.py is the updated version of Lab06.py. It uses the WAQI
API to get the AQI index.